Added more proper messageSizes where they were still 1000, specifically on the texture parts. Hopefully it'll still work properly for you.
Also stepped through the program a lot to find the issue my computer has with it. At occasion (3 times today) I get the alpha texture to show up in the RTV. The first time was via hardcoding which showed it blue as it should be, but this refused to work later on. The other two it was through picking it as the file texture and stepping through the program slowly, followed by immediate crash if an attempt to add further materials afterward was made.
Meanwhile textures that aren't alpha.dds seem to not alert the callback and so not send any message right now.